course-service/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
PyJWT==2.8.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
